f: modify rem to 10px, moved styleConstants to root client directory
