+++ client/src/component/Content/styles.tsx
@@ -1,13 +1,14 @@
 import styled from 'styled-components/macro';
 
 export const Content = styled.div`
-  margin-top: 4rem;
+  margin-top: 6rem;
 `;
 export const Container = styled.div`
-  width: 64rem;
+  min-width: 48rem;
   height: 20rem;
-  margin: 0px auto;
+  margin: 0 auto;
   display: flex;
   justify-content: center;
   align-items: center;
+  background-color: beige;
 `;
