--- client/src/component/GlobalStyle/styleConstants.ts
@@ -1,12 +1,0 @@
-const COLORS = {
-  fontDark: '#3c4146',
-  fontLight: '#ffffff',
-  primary: '#f94e36',
-  primaryLight: '#ff8262',
-  primaryDark: '#be070b ',
-  secondary: '#ffb82b',
-  secondaryLight: '#ffea61',
-  secondaryDark: '#c78800',
-};
-
-export default COLORS;
